refactor: change NavigationModifier location
